[processing] rename some algorithms to avoid naming conflicts and maintain consistency
