+++ python/plugins/processing/tests/testdata/gdal_algorithm_tests.yaml
@@ -1,4 +1,5 @@
 tests:
+
   - algorithm: gdal:gdalinfo
     name: gdalinfo
     params:
@@ -273,7 +274,39 @@ tests:
         hash: a497763f44380ec33acab572a926a6516e875f2284fc55e702abbdfc
         type: rasterhash
 
-#  - algorithm: gdal:ogrinfo
+#  - algorithm: gdal:merge
+#    name: Merge 2 raster layers
+#    params:
+#      INPUT:
+#        params:
+#        - name: custom/dem_to_merge1.tif
+#          type: raster
+#        - name: custom/dem_to_merge2.tif
+#          type: raster
+#        type: multi
+#      PCT: false
+#      RTYPE: '5'
+#      SEPARATE: false
+#    results:
+#      OUTPUT:
+#        hash: fb6f0f4081308cc188eda2fbe106e04e0ebf866fc56853f0cca8db1c
+#        type: rasterhash
+#
+#  - algorithm: gdal:rastercalculator
+#    name: Simple Raster Calculator test with only 1 raster
+#    params:
+#      BAND_A: '1'
+#      FORMULA: A/2
+#      INPUT_A:
+#        name: dem.tif
+#        type: raster
+#      RTYPE: '5'
+#    results:
+#      OUTPUT:
+#        hash: 2d0585dc8166fbe3cec7d9e6fd66e95cc9bdb1043b3b0fa7cbdfef4c
+#        type: rasterhash
+
+# - algorithm: gdal:ogrinfo
 #    name: ogrinfo
 #    params:
 #      INPUT:
@@ -288,6 +321,57 @@ tests:
 #          - 'Extent: \(-1.000000, -3.000000\) - \(11.000000, 5.000000\)'
 #          - 'Geometry: Line String'
 #          - 'Feature Count: [6|7]' # On some platforms returns 6 instead of 7...
+
+  - algorithm: gdal:pointsalonglines
+    name: Points along lines
+    params:
+      DISTANCE: 0.25
+      GEOMETRY: geometry
+      INPUT:
+        name: lines.gml
+        type: vector
+    results:
+      OUTPUT:
+        name: expected/gdal/points_along_lines.gml
+        type: vector
+        pk: fid
+        compare:
+          ignore_crs_check: true
+
+#  - algorithm: gdal:offsetlinesforlines
+#    name: Offset lines for lines (right-handed)
+#    params:
+#      GEOMETRY: geometry
+#      INPUT_LAYER:
+#         name: lines.gml
+#         type: vector
+#      RADIUS: -0.5
+#    results:
+#      OUTPUT_LAYER:
+#        name: expected/gdal/offset_lines.gml
+#        type: vector
+#        compare:
+#          geometry:
+#            precision: 7
+#
+#  - algorithm: gdal:singlesidedbufferforlines
+#    name: One-side buffer for lines (left-handed)
+#    params:
+#      DISSOLVEALL: false
+#      GEOMETRY: geometry
+#      INPUT_LAYER:
+#        name: lines.gml
+#        type: vector
+#      LEFTRIGHT: '1'
+#      MULTI: false
+#      RADIUS: 0.5
+#    results:
+#      OUTPUT_LAYER:
+#        name: expected/gdal/one_side_buffer.gml
+#        type: vector
+#        compare:
+#          geometry:
+#            precision: 7
 #
 #  - algorithm: gdal:buffervectors
 #    name: Buffer lines
@@ -345,103 +429,4 @@ tests:
 #          geometry:
 #            precision: 7
 #
-  - algorithm: gdal:createpointsalonglines
-    name: Points along lines
-    params:
-      DISTANCE: 0.25
-      GEOMETRY: geometry
-      INPUT:
-        name: lines.gml
-        type: vector
-    results:
-      OUTPUT:
-        name: expected/gdal/points_along_lines.gml
-        type: vector
-        pk: fid
-        compare:
-          ignore_crs_check: true
 
-#  - algorithm: gdal:offsetlinesforlines
-#    name: Offset lines for lines (right-handed)
-#    params:
-#      GEOMETRY: geometry
-#      INPUT_LAYER:
-#         name: lines.gml
-#         type: vector
-#      RADIUS: -0.5
-#    results:
-#      OUTPUT_LAYER:
-#        name: expected/gdal/offset_lines.gml
-#        type: vector
-#        compare:
-#          geometry:
-#            precision: 7
-#
-#  - algorithm: gdal:singlesidedbufferforlines
-#    name: One-side buffer for lines (left-handed)
-#    params:
-#      DISSOLVEALL: false
-#      GEOMETRY: geometry
-#      INPUT_LAYER:
-#        name: lines.gml
-#        type: vector
-#      LEFTRIGHT: '1'
-#      MULTI: false
-#      RADIUS: 0.5
-#    results:
-#      OUTPUT_LAYER:
-#        name: expected/gdal/one_side_buffer.gml
-#        type: vector
-#        compare:
-#          geometry:
-#            precision: 7
-#
-#  - algorithm: gdal:merge
-#    name: Merge 2 raster layers
-#    params:
-#      INPUT:
-#        params:
-#        - name: custom/dem_to_merge1.tif
-#          type: raster
-#        - name: custom/dem_to_merge2.tif
-#          type: raster
-#        type: multi
-#      PCT: false
-#      RTYPE: '5'
-#      SEPARATE: false
-#    results:
-#      OUTPUT:
-#        hash: fb6f0f4081308cc188eda2fbe106e04e0ebf866fc56853f0cca8db1c
-#        type: rasterhash
-#
-#  - algorithm: gdal:merge
-#    name: Merge layers and separate in bands
-#    params:
-#      INPUT:
-#        params:
-#        - name: custom/dem_to_merge1.tif
-#          type: raster
-#        - name: custom/dem_to_merge2.tif
-#          type: raster
-#        type: multi
-#      PCT: false
-#      RTYPE: '5'
-#      SEPARATE: true
-#    results:
-#      OUTPUT:
-#        hash: cbf7b8ca91e7d6cab5ef76eeed6ce80e5dbe173afb7b8eb4aed2e969
-#        type: rasterhash
-#
-#  - algorithm: gdal:rastercalculator
-#    name: Simple Raster Calculator test with only 1 raster
-#    params:
-#      BAND_A: '1'
-#      FORMULA: A/2
-#      INPUT_A:
-#        name: dem.tif
-#        type: raster
-#      RTYPE: '5'
-#    results:
-#      OUTPUT:
-#        hash: 2d0585dc8166fbe3cec7d9e6fd66e95cc9bdb1043b3b0fa7cbdfef4c
-#        type: rasterhash
